Introduce support for different winding rules
This is a model which can support all OpenGL winding rules as well as
any other rule which is just a test of the winding number (if there are
any such tests that are useful).

This requires introducing a flag on Segment objects to track whether they
have been "swapped": A Segment which is not swapped has its endpoints in
the proper order to be consecutive points in a CCW-wound polygon.  One
which is swapped is in the opposite order.

It is necessary to invert this flag when transforming a segment by a
transformation with a negative determinant, and when swapping the endpoints
to satisfy a sorting criterion.